# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AnthoLume API v1
  version: 1.0.0
  description: REST API for AnthoLume document management system
servers:
  - url: /api/v1
components:
  schemas:
...
    DocumentsResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        next_page:
          type: integer
          format: int64
        previous_page:
          type: integer
          format: int64
        search:
          type: string
      required:
        - documents
        - total
        - page
        - limit
    DocumentResponse:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/Document'
      required:
        - document
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    DatabaseInfo:
      type: object
      properties:
        documents_size:
          type: integer
          format: int64
        activity_size:
          type: integer
          format: int64
        progress_size:
          type: integer
          format: int64
        devices_size:
          type: integer
          format: int64
      required:
        - documents_size
        - activity_size
        - progress_size
        - devices_size
...
    BackupType:
      type: string
      enum: [COVERS, DOCUMENTS]
...
    OperationType:
      type: string
      enum: [CREATE, UPDATE, DELETE]
    User:
      type: object
      properties:
        id:
          type: string
        admin:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - admin
        - created_at
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: cookie
      name: token
paths:
  /documents:
    get:
      summary: List documents
      operationId: getDocuments
      tags:
        - Documents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int64
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            default: 9
        - name: search
          in: query
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Upload a new document
      operationId: createDocument
      tags:
        - Documents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document_file:
                  type: string
                  format: binary
              required:
                - document_file
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}:
    get:
      summary: Get a single document
      operationId: getDocument
      tags:
        - Documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Update document editable fields
      operationId: editDocument
      tags:
        - Documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                description:
                  type: string
                isbn10:
                  type: string
                isbn13:
                  type: string
                cover_gbid:
                  type: string
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin:
    get:
      summary: Get admin page data
      operationId: getAdmin
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  database_info:
                    $ref: '#/components/schemas/DatabaseInfo'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Perform admin action (backup, restore, etc.)
      operationId: postAdminAction
      tags:
        - Admin
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [BACKUP, RESTORE, METADATA_MATCH, CACHE_TABLES]
                backup_types:
                  type: array
                  items:
                    $ref: '#/components/schemas/BackupType'
                restore_file:
                  type: string
                  format: binary
              required:
                - action
      security:
        - BearerAuth: []
      responses:
        200:
          description: Action completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      summary: Get all users
      operationId: getUsers
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create, update, or delete user
      operationId: updateUser
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                operation:
                  $ref: '#/components/schemas/OperationType'
                user:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
              required:
                - operation
                - user
      security:
        - BearerAuth: []
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...